backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlalchemy
pydantic
requests
sentence-transformers
faiss-cpu
numpy